[UHYU-248] refactor: 기존 userStore 삭제 및 수정
